Daother 8061

+++ deploy/launch.sh
@@ -13,12 +13,18 @@ export APP_MODULE="rcsb.app.${SERVICE_NAME}.main:app"
 export GUNICORN_CONF=${GUNICORN_CONF:-"/app/gunicorn_conf.py"}
 export WORKER_CLASS=${WORKER_CLASS:-"uvicorn.workers.UvicornWorker"}
 
-HERE="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-TOPDIR="$(dirname "$HERE")"
-echo "HERE=${HERE}"
-echo "TOPDIR=${TOPDIR}"
+THISFILE=`basename $0`
+if [ $THISFILE != 'launch.sh' ]
+then
+  echo "error in launch.sh - \$0 does not produce correct result on this system"
+  exit 1
+fi
+HERE=`cd $(dirname $0); pwd`
+TOPDIR=`dirname $HERE`
+echo "HERE=$HERE"
+echo "TOPDIR=$TOPDIR"
 UPTIME_START=`echo $(date +%s)`
-echo "UPTIME_START=${UPTIME_START}"
+echo "UPTIME_START=$UPTIME_START"
 echo $UPTIME_START > uptime.txt
 
 # Optional setup.sh
